requirements.txt
asgiref==3.8.1
boto3 @ file:///storage/emulated/0/HCA/boto3-1.35.86.tar.gz#sha256=d61476fdd5a5388503b72c897083310d2329ce088593c4332b571a860be5d155
botocore==1.35.86
certifi==2024.12.14
charset-normalizer==3.4.0
distlib==0.3.9
dj-database-url==2.3.0
Django==4.2.17
django-allauth==65.3.0
django-crispy-forms==2.3
django-jazzmin==3.0.1
django-jet==1.0.8
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-phonenumber-field==8.0.0
django-storages==1.14.4
django-suit==0.2.28
django_notifications==0.1.dev0
filelock==3.16.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
jsonfield==3.1.0
mailjet-rest==1.3.4
packaging==24.2
phonenumbers==8.13.52
pillow==11.0.0
platformdirs==4.3.6
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasynchat==1.0.4
pyasyncore==1.0.4
pyftpdlib==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
sendgrid==3.6.5
sendgrid-django==4.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
swapper==1.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
wheel==0.45.1
whitenoise==6.8.2
